tests/test_spark_processing.py
import os
import sys
from pathlib import Path
from dotenv import load_dotenv

# Add project root to Python path
project_root = Path(__file__).parent
sys.path.append(str(project_root))

# Load environment variables
load_dotenv()

from spark_app.spark_streaming_processor import main

if __name__ == "__main__":
    main()