requirements.txt
# Core dependencies
pyspark==3.3.2
kafka-python==2.0.2
pymongo>=3.6.0,<4.0.0
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0

# Data processing
pandas>=1.5.3,<2.0.0
numpy>=1.24.2,<2.0.0
scipy>=1.10.1,<2.0.0

# Airflow dependencies
apache-airflow==2.5.1
apache-airflow-providers-mongo==3.1.1
apache-airflow-providers-postgres==5.4.0
colorlog>=4.0.2,<5.0.0

# Dashboard dependencies
flask>=2.2.3,<3.0.0
flask-cors>=3.0.10,<4.0.0
plotly>=5.13.1,<6.0.0

# Testing dependencies
pytest>=7.2.2,<8.0.0
pytest-mock>=3.10.0,<4.0.0

# Utilities
schedule>=1.1.0,<2.0.0